requirements.txt
ansible==13.5.0
ansible-lint==26.4.0
# https://github.com/pyca/bcrypt/issues/684
bcrypt==5.0.0
jmespath==1.1.0
netaddr==1.3.0
openshift==0.13.2
passlib==1.7.4
PyYAML==6.0.2